refactor(app/consumer): merge debt gate into MsgFilterDecorator
ConsumerFundsDecorator and MsgFilterDecorator were both consumer-state
admission gates differing only in phase: MsgFilter handled the pre-CCV
window (only /ibc.* allowed), ConsumerFunds handled the in-debt window
(only /ibc.core.* and /cosmos.gov.* allowed). Folding them together
expresses the three modes (pre-CCV / normal / in-debt) as one switch
and drops a redundant GetProviderClientID call per tx in the hot path.
